backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openpyxl
redis
aioredis
fakeredis
reportlab
pandas
numpy
